fixed reference errors

+++ migrations/001.do.create_poker_winrate_sessions.sql
@@ -6,10 +6,10 @@ CREATE TABLE poker_sessions (
   game_type_one game_type_one NOT NULL,
   game_type_two game_type_two NOT NULL,
   date_played TIMESTAMP DEFAULT now() NOT NULL,
-  small_blind INTEGER NOT NULL,
-  big_blind INTEGER NOT NULL,
-  buy_in INTEGER NOT NULL,
-  cashed_out INTEGER NOT NULL,
-  session_length INTEGER NOT NULL,
+  small_blind NUMERIC NOT NULL,
+  big_blind NUMERIC NOT NULL,
+  buy_in NUMERIC NOT NULL,
+  cashed_out NUMERIC NOT NULL,
+  session_length NUMERIC NOT NULL,
   notes TEXT
 );
